Migrate the test suite to the reworked record encoding
The record-encoding rework left the test target uncompilable. Update the
test helpers and suites to the current Sources APIs: plain string record
identifiers (RecordID removed), Record(recordType:recordID:), RecordQuery
filters/sorts in place of HTTPFilter/HTTPSort, the recordType-less Matrix
initializer, and the per-backend SyncDelivery model. Restore client-side
activity/metric-series reconstruction as a shared test helper and drop
tests for APIs that no longer exist.

+++ Tests/ScoutTests/Core/Backend/DeliverTests.swift
@@ -167,7 +167,7 @@ struct DeliverTests {
         let event = EventObject.stub(name: "login", synced: true, in: context)
         event.seedDelivery([.raw, .matrix], for: "cloud", in: context)
         // The server is already at the attempt ceiling and still owes its raw record.
-        event.seedDelivery([.raw], attempts: SyncDelivery.maxAttempts, for: "server", in: context)
+        event.seedDelivery([.raw], attempts: Int16(SyncDelivery.maxAttempts), for: "server", in: context)
         try context.save()
 
         try await deliver(EventObject.self, to: cloudBackend)
